Made a couple of routes for Spots

+++ backend/db/models/spot.js
@@ -21,8 +21,8 @@ module.exports = (sequelize, DataTypes) => {
     city: DataTypes.STRING,
     state: DataTypes.STRING,
     country: DataTypes.STRING,
-    lat: DataTypes.INTEGER,
-    lng: DataTypes.INTEGER,
+    lat: DataTypes.FLOAT,
+    lng: DataTypes.FLOAT,
     name: {
       type: DataTypes.STRING,
       allowNull: false,
